Refine canvas selection and update grid layout controls
This commit removes obsolete x position variable (mx), fixes visibility and origin issues with the canvas selection, and introduces changes in grid layout controls. Grid layout controls are now set to not autosize and initialise as invisible. Further, origin adjustments are made based on the visibility state of canvas selection in different scenarios.